requirements.txt
alembic==1.5.8
aniso8601==9.0.1
appdirs==1.4.4
attrs==20.3.0
autoenv==1.0.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.7.0
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
pbr==5.5.1
pycparser==2.20
PyJWT==2.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.4.5
stevedore==3.3.0
urllib3==1.26.2
virtualenv==20.4.3
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
Werkzeug==0.16.1
